requirement.txt

charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
Flask==2.1.3
fonttools==4.34.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
pytz==2022.1
pywin32==302
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.8.1
seaborn==0.11.2
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
urllib3==1.26.10
webencodings==0.5.1
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.1
